+++ app/src/main/java/org/breezyweather/common/source/LocationPreset.kt
@@ -60,7 +60,6 @@ enum class LocationPreset(
     ITALY("meteoam", airQuality = "openmeteo", pollen = "openmeteo", minutely = "openmeteo", normals = "accu"),
     NORWAY("metno", pollen = "openmeteo", alert = "accu", normals = "accu"),
     SWEDEN("smhi", airQuality = "openmeteo", pollen = "openmeteo", minutely = "metno", alert = "accu", normals = "accu"),
-    UNITED_KINGDOM("metoffice", current = "openmeteo", airQuality = "openmeteo", pollen = "openmeteo", minutely = "openmeteo", alert = "accu", normals = "accu"),
 
     // Asia
     // Do NOT set up other sources as only 中国 source is not rate-limited by the Great Firewall
@@ -84,7 +83,6 @@ enum class LocationPreset(
                     "DK" -> DENMARK
                     "FI" -> FINLAND
                     "FR" -> FRANCE
-                    "GB" -> UNITED_KINGDOM
                     "IE" -> IRELAND
                     "IT", "SM", "VA" -> ITALY
                     "NO" -> NORWAY